Added more global constants and variables to the main LFEDocumentClassifier script to allow better control over the pipeline.
Added function to generate a bag of words from any
current pipeline methods.

Added function to create function masks using a bag
of words and any scored word pairs.

Renamed BagOfWords file to FeatureCreation to cover
a wider range of responsibilities.

Changed FeatureCreation to be a script of functions
rather than a class as this fits the structure more
appropriately.

Restructured the TextRank class to better fit OO standards