Added threshold option when generating features from keywords etc. Allows only the top x keywords to be include in the bag of words. Off by default.
Added statistics and results generator class for ordering/colating results, performing
simple stats tasks and (in the future) file IO for saving these.

Refactored text rank to use NLTK inplace of spaCy NLP for tokenisation of words/sentances.
This improved runtime of this method from 23 seconds to 2 seconds. The old method has been
kept as "DEPRECIATED" and can be used if POS tagging etc from spaCy NLP is needed in the
future.