Combined all parameter (used to be names constant) files into one large file (ProgramParameters.py) for easier access to all parameters.
Added a raw words option to perform a basic Bag of Words approach using
only term frequency (with option to expand for TF-IDF in the future).

Added extra parameters for further control of the pipeline (bag of words
parameters and extra preprocessing parameters).

Refactored some of the text rank methods to be pre processing static
methods so that they could be used in other pipe lines.