Added new Support Vector Machine (SVM) classifier.
Added required parameters for adjustment of the SVM classifier into
the program parameters.

Updated the pipeline to allow for SVM to be incorporated.

Added multi-label classification option to SVM, and the main pipeline.

Made minor refactoring changes to name protocol in program parameters.

Made minor refactoring changes to layout of main pipeline script.